Encode defaults in serialization for the necessary classes
As this indicates
https://github.com/Kotlin/kotlinx.serialization/blob/master/docs/basic-serialization.md#defaults-are-not-encoded-by-default
Default values are not encoded by default since it's considered that on
deserialization it will then be auto-generate if the value was the
default one anyway. The backend of course can not make such assumptions
and wants to see the value, so we simply enforce those scenarios to
always encode those values.